move enums into package
