+++ src/main/java/com/develcode/develfood/model/CustomerFavoritePlate.java
@@ -1,5 +1,6 @@
 package com.develcode.develfood.model;
 
+import com.develcode.develfood.model.enums.Plate;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
