+++ src/main/java/com/develcode/develfood/model/RequestItem.java
@@ -1,5 +1,6 @@
 package com.develcode.develfood.model;
 
+import com.develcode.develfood.model.enums.Plate;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -39,7 +40,4 @@ public class RequestItem {
 
     private String observation;
 
-    public void setRequest(Request request) {
-        this.request = request;
-    }
 }
